requirements.txt
pypdf
langchain_community
langchain
streamlit
langchain-text-splitters
langchain_openai
python-dotenv
pinecone-client
langchain_pinecone
streamlit_chat